+++ examples/subscription/server.go
@@ -24,7 +24,7 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/BoltApp/braintree-go"
+	"github.com/niharika88/braintree-go"
 )
 
 func main() {